python_code/requirements.txt
pynput==1.4.2
pyserial==3.4
